undo previous bogas bugfix; it was a stupid bracket

+++ reportFlow.py
@@ -15,7 +15,7 @@ def generateReport(analysisResults, params, logFilePath, writeToDisk):
 	fullProteinDF = analysisResults[1]
 	PCAResult = analysisResults[2]
 	HCResult = analysisResults[3]
-	allExperimentsIntensitiesPerCommonPeptide = np.asarray(analysisResults[4], dtype=float)
+	allExperimentsIntensitiesPerCommonPeptide = np.asarray(analysisResults[4])
 	metadata = analysisResults[5]
 
 	### TEST
